+++ src/components/markdown/ChangeLog.md
@@ -3,6 +3,7 @@
 ### 25 September 2021
 
 * Added creature sprites, sourced from the Siralim Ultimate API.
+* Tool now allows you to upload your build directly from the export from the game.
 
 ### 24 September 2021